# openapi.yaml
openapi: 3.0.3
info:
  title: 唐宋电竞陪玩报单平台API
  description: 唐宋电竞陪玩报单平台后端API文档
  version: 1.0.0
  contact:
    name: 唐宋电竞技术团队
servers:
  - url: http://localhost:8080/api/v1
    description: 开发环境
  - url: https://api.tangsong-esports.com/api/v1
    description: 生产环境
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
servers:
  - url: http://localhost:8080/api/v1
    description: 开发环境
  - url: https://api.tangsong-esports.com/api/v1
    description: 生产环境
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "操作成功"
        data:
          type: object
    PageResponse:
      type: object
      properties:
        list:
          type: array
          items:
            type: object
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 10

    # 登录相关
...
    Member:
      type: object
      properties:
        member_id:
          type: integer
          example: 1
        account:
          type: string
          example: "admin"
        name:
          type: string
          example: "系统管理员"
        phone_number:
          type: string
          example: "13800138000"
        department:
          type: string
          example: "管理部"
        user_role:
          type: string
          example: "超级管理员"
        status:
          type: string
          enum: ["正常", "禁用"]
          example: "正常"
        is_enabled:
          type: boolean
          example: true
        notes:
          type: string
          example: "系统管理员账号"
        last_login_at:
          type: string
          format: date-time
        last_login_ip:
          type: string
          example: "127.0.0.1"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        permissions:
          $ref: '#/components/schemas/MemberPermissions'
        financial_settings:
          $ref: '#/components/schemas/MemberFinancialSettings'
    MemberPermissions:
      type: object
      properties:
        permission_id:
          type: integer
        member_id:
          type: integer
        is_auditor:
          type: boolean
          example: true
        can_report:
          type: boolean
          example: true
        can_accept_order:
          type: boolean
          example: true
    MemberFinancialSettings:
      type: object
      properties:
        setting_id:
          type: integer
        member_id:
          type: integer
        commission_rate:
          type: number
          format: decimal
          example: 0.15
    MemberCreateRequest:
      type: object
      required:
        - account
        - password
        - name
        - user_role
      properties:
        account:
          type: string
          example: "test001"
        password:
          type: string
          example: "123456"
        name:
          type: string
          example: "测试陪玩"
        phone_number:
          type: string
          example: "13800138000"
        department:
          type: string
          example: "陪玩部"
        user_role:
          type: string
          example: "陪玩"
        notes:
          type: string
          example: "测试账号"
        is_auditor:
          type: boolean
          default: false
        can_report:
          type: boolean
          default: true
        can_accept_order:
          type: boolean
          default: true
        commission_rate:
          type: number
          format: decimal
          example: 0.15
        creator_id:
          type: integer
          example: 1
        assignee_id:
          type: integer
          example: 1

    # 客户相关
...
paths:
...
  /members:
    get:
      tags:
        - 成员管理
      summary: 获取成员列表
      description: 分页获取内部成员列表，支持搜索和筛选
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          description: 每页数量
          schema:
            type: integer
            default: 10
        - name: account
          in: query
          description: 账号搜索
          schema:
            type: string
        - name: name
          in: query
          description: 姓名搜索
          schema:
            type: string
        - name: department
          in: query
          description: 部门筛选
          schema:
            type: string
        - name: status
          in: query
          description: 状态筛选
          schema:
            type: string
            enum: ["正常", "禁用"]
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PageResponse'

    post:
      tags:
        - 成员管理
      summary: 创建成员
      description: 创建新的内部成员
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberCreateRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Member'
  /members/{id}:
    get:
      tags:
        - 成员管理
      summary: 获取成员详情
      description: 根据ID获取内部成员详细信息
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 成员ID
          schema:
            type: integer
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Member'

    put:
      tags:
        - 成员管理
      summary: 更新成员
      description: 更新内部成员信息
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 成员ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberCreateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Member'

    delete:
      tags:
        - 成员管理
      summary: 删除成员
      description: 软删除内部成员
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 成员ID
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...